refactor: split patient service into different files
