+++ patients/services/__init__.py
@@ -1,0 +1,1 @@
+from .patient_service import get_patient_service, patient_service
