+++ patients/services/api_patient_service.py
@@ -1,112 +1,22 @@
-from uuid import UUID
 import requests
-from abc import ABC, abstractmethod
 from typing import List
 import logging
 from django.core.cache import cache
 
 from larvixon_site.settings import (
-    MOCK_PATIENT_SERVICE,
     PATIENT_API_TOKEN,
-    PATIENT_SERVICE_URL,
 )
 from patients.errors import (
-    PatientInvalidUUIDError,
     PatientServiceUnavailableError,
     PatientServiceResponseError,
 )
+from patients.services.base_patient_service import BasePatientService
 
 logger: logging.Logger = logging.getLogger(__name__)
 
 PESEL_ID = "http://hl7.org/fhir/sid/pesel"
 TIMEOUT_SECONDS = 90
 CACHE_TIME_SECONDS = 60
-
-
-class BasePatientService(ABC):
-    @abstractmethod
-    def search_patients(
-        self,
-        first_name: str | None = None,
-        last_name: str | None = None,
-        pesel: str | None = None,
-    ) -> List[dict]:
-        """Search for patients by first name, last name, or PESEL."""
-        pass
-
-    @abstractmethod
-    def get_patient_by_guid(self, guid: str) -> dict | None:
-        """Get a single patient by their GUID."""
-        pass
-
-    @abstractmethod
-    def get_patients_by_guids(self, guids: List[str]) -> dict[str, dict]:
-        """Get multiple patients by their GUIDs."""
-        pass
-
-    def validate_uuid(self, guid: str) -> None:
-        """Validate that the provided GUID is a valid UUID."""
-        try:
-            UUID(guid)
-        except ValueError:
-            logger.warning(f"Invalid Patient GUID format provided: {guid}")
-            raise PatientInvalidUUIDError(f"Invalid Patient GUID format: {guid}")
-
-
-class MockPatientService(BasePatientService):
-    def _get_mock_patient(self) -> dict:
-        return {
-            "id": "00000000-0000-0000-0000-000000000001",
-            "pesel": "90010112345",
-            "first_name": "Jan",
-            "last_name": "Kowalski",
-            "birth_date": "1990-01-01",
-            "gender": "male",
-            "phone": "+48123456789",
-            "email": "jan.kowalski@example.com",
-            "address_line": "ul. Przykładowa 1",
-            "city": "Warszawa",
-            "postal_code": "00-001",
-            "country": "PL",
-        }
-
-    def search_patients(
-        self,
-        first_name: str | None = None,
-        last_name: str | None = None,
-        pesel: str | None = None,
-    ) -> List[dict]:
-        mock_patient = self._get_mock_patient()
-
-        if first_name or last_name or pesel:
-            if (
-                first_name
-                and first_name.lower() in mock_patient["first_name"].lower()
-                or last_name
-                and last_name.lower() in mock_patient["last_name"].lower()
-                or pesel
-                and pesel in mock_patient["pesel"]
-            ):
-                return [mock_patient]
-            return []
-
-        return [mock_patient]
-
-    def get_patient_by_guid(self, guid: str) -> dict | None:
-        mock_patient = self._get_mock_patient()
-        if mock_patient["id"] == guid:
-            return mock_patient
-        return None
-
-    def get_patients_by_guids(self, guids: List[str]) -> dict[str, dict]:
-        mock_patient = self._get_mock_patient()
-        results = {}
-
-        for guid in guids:
-            if mock_patient["id"] == guid:
-                results[guid] = mock_patient
-
-        return results
 
 
 class APIPatientService(BasePatientService):
@@ -318,12 +228,3 @@ class APIPatientService(BasePatientService):
             "postal_code": postal_code,
             "country": country,
         }
-
-
-def get_patient_service() -> BasePatientService:
-    if MOCK_PATIENT_SERVICE:
-        return MockPatientService()
-    return APIPatientService(PATIENT_SERVICE_URL)
-
-
-patient_service: BasePatientService = get_patient_service()
