+++ patients/tests/test_patient_views.py
@@ -349,7 +349,7 @@ class TestPatientViews(APITestCase):
         User.objects.all().delete()
         cache.clear()
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_get_patient_success(self, mock_get: Mock) -> None:
         """Test successful patient retrieval with valid GUID."""
         guid = "ab758f9b-0298-4823-b144-ae0db20bc215"
@@ -374,7 +374,7 @@ class TestPatientViews(APITestCase):
         self.assertEqual(response.data["first_name"], "Aurelia")
         self.assertEqual(response.data["last_name"], "Jędruszczak")
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_get_patient_not_found(self, mock_get: Mock) -> None:
         """Test patient retrieval with non-existent GUID."""
         guid = "99999999-9999-9999-9999-999999999999"
@@ -394,7 +394,7 @@ class TestPatientViews(APITestCase):
         self.assertIn("detail", response.data)
         self.assertEqual(response.data["detail"], "Patient not found.")
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_get_patient_invalid_guid_format(self, mock_get: Mock) -> None:
         """Test patient retrieval with malformed GUID."""
         invalid_guids = [
@@ -435,7 +435,7 @@ class TestPatientViews(APITestCase):
 
         self.assertEqual(response.status_code, 401)
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_search_patients_success(self, mock_get: Mock) -> None:
         """Test successful patient search."""
         mock_response = Mock()
@@ -452,7 +452,7 @@ class TestPatientViews(APITestCase):
         self.assertIsInstance(response.data, list)
         self.assertEqual(len(response.data), 6)
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_search_patients_with_search_term(self, mock_get: Mock) -> None:
         """Test patient search with search term."""
         search_term = "Aurelia"
@@ -490,7 +490,7 @@ class TestPatientViews(APITestCase):
             or search_term.lower() in patient["last_name"].lower()
         )
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_search_patients_no_results(self, mock_get: Mock) -> None:
         """Test patient search with search term that has no matches."""
         search_term = "NonExistentPatient12345"
@@ -516,7 +516,7 @@ class TestPatientViews(APITestCase):
         self.assertIsInstance(response.data, list)
         self.assertEqual(len(response.data), 0)
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_search_patients_empty_search_term(self, mock_get: Mock) -> None:
         """Test patient search with empty search term."""
         mock_response = Mock()
@@ -532,7 +532,7 @@ class TestPatientViews(APITestCase):
         self.assertEqual(response.status_code, 200)
         self.assertIsInstance(response.data, list)
 
-    @patch("patients.services.patients.requests.get")
+    @patch("patients.services.api_patient_service.requests.get")
     def test_search_patients_special_characters(self, mock_get: Mock) -> None:
         """Test patient search with special characters in search term."""
         special_chars = ["%", "&", "ą", "ł", "ż"]
