+++ videoprocessor/tests/test_upload_views.py
@@ -1,11 +1,12 @@
-from unittest.mock import patch
+from unittest.mock import patch, MagicMock
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.core.files.base import ContentFile
 from django.urls import reverse
+from django.test import override_settings
 from rest_framework.test import APIRequestFactory, force_authenticate, APITestCase
 from rest_framework import status
 from accounts.models import User
-from analysis.models import VideoAnalysis
+from analysis.models import VideoAnalysis, Substance
 from videoprocessor.views.upload import VideoUploadView
 from tests.common import TestFixtures, cleanup_test_media
 
@@ -42,13 +43,22 @@ class TestVideoUploadView(APITestCase):
             VIDEO_FILENAME, VIDEO_CONTENT, content_type=VIDEO_CONTENT_TYPE
         )
 
+    @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
     @patch(
         "videoprocessor.services.video_file_manager.VideoFileManager.extract_and_save_first_frame"
     )
-    @patch("videoprocessor.tasks.process_video_task.delay")
-    def test_upload_with_default_video_manager(self, mock_task, mock_extract):
-        """Should work with default VideoFileManager."""
+    @patch("requests.post")
+    def test_upload_with_default_video_manager(self, mock_requests_post, mock_extract):
+        """Should work with default VideoFileManager and make request to ML endpoint."""
         mock_extract.return_value = ("test_thumb.jpg", ContentFile(b"fake thumbnail"))
+
+        # Mock the ML endpoint response
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.json.return_value = {
+            "predictions": {"cocaine": 85.5, "morphine": 10.2, "ethanol": 4.3}
+        }
+        mock_requests_post.return_value = mock_response
 
         video_file = self._create_video_file()
 
@@ -63,17 +73,28 @@ class TestVideoUploadView(APITestCase):
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         self.assertIn("analysis_id", response.data)
+        # Verify ML endpoint was called
+        mock_requests_post.assert_called_once()
 
+    @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
     @patch(
         "videoprocessor.services.video_file_manager.VideoFileManager.extract_and_save_first_frame"
     )
-    @patch("videoprocessor.tasks.process_video_task.delay")
-    def test_upload_with_injected_video_manager(self, mock_task, mock_extract):
-        """Should accept injected VideoFileManager for testing."""
+    @patch("requests.post")
+    def test_upload_with_injected_video_manager(self, mock_requests_post, mock_extract):
+        """Should accept injected VideoFileManager and process with real ML service."""
         mock_extract.return_value = (
             "test_thumb.jpg",
             ContentFile(b"fake thumbnail"),
         )
+
+        # Mock the ML endpoint response
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.json.return_value = {
+            "predictions": {"cocaine": 85.5, "morphine": 10.2}
+        }
+        mock_requests_post.return_value = mock_response
 
         video_file = self._create_video_file()
 
@@ -88,6 +109,7 @@ class TestVideoUploadView(APITestCase):
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         mock_extract.assert_called_once()
+        mock_requests_post.assert_called_once()
 
     def test_upload_without_video_file(self):
         """Should return 400 when no video file provided."""
@@ -123,15 +145,16 @@ class TestVideoUploadView(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertIn("Invalid Patient GUID", response.data["error"])
 
+    @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
     @patch(
         "videoprocessor.services.video_file_manager.VideoFileManager.extract_and_save_first_frame"
     )
     @patch(
         "videoprocessor.services.video_upload_service.patient_service.get_patient_by_guid"
     )
-    @patch("videoprocessor.tasks.process_video_task.delay")
+    @patch("requests.post")
     def test_upload_with_valid_patient_guid(
-        self, mock_task, mock_get_patient, mock_extract
+        self, mock_requests_post, mock_get_patient, mock_extract
     ):
         """Should accept valid patient GUID."""
         mock_extract.return_value = ("test_thumb.jpg", ContentFile(b"fake thumbnail"))
@@ -140,6 +163,14 @@ class TestVideoUploadView(APITestCase):
             "first_name": "John",
             "last_name": "Doe",
         }
+
+        # Mock the ML endpoint response
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.json.return_value = {
+            "predictions": {"cocaine": 85.5, "morphine": 10.2}
+        }
+        mock_requests_post.return_value = mock_response
 
         video_file = self._create_video_file()
 
@@ -158,6 +189,7 @@ class TestVideoUploadView(APITestCase):
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         mock_get_patient.assert_called_once_with("00000000-0000-0000-0000-000000000001")
+        mock_requests_post.assert_called_once()
 
     @patch(
         "videoprocessor.services.video_upload_service.patient_service.get_patient_by_guid"
@@ -226,13 +258,22 @@ class TestVideoUploadView(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertIn("Unsupported file format", response.data["error"])
 
+    @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
     @patch(
         "videoprocessor.services.video_file_manager.VideoFileManager.extract_and_save_first_frame"
     )
-    @patch("videoprocessor.tasks.process_video_task.delay")
-    def test_upload_triggers_background_task(self, mock_task, mock_extract):
-        """Should trigger process_video_task after successful upload."""
+    @patch("requests.post")
+    def test_upload_triggers_background_task(self, mock_requests_post, mock_extract):
+        """Should process video with ML service after successful upload."""
         mock_extract.return_value = ("test_thumb.jpg", ContentFile(b"fake thumbnail"))
+
+        # Mock the ML endpoint response
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.json.return_value = {
+            "predictions": {"cocaine": 85.5, "morphine": 10.2}
+        }
+        mock_requests_post.return_value = mock_response
 
         video_file = self._create_video_file()
 
@@ -246,7 +287,75 @@ class TestVideoUploadView(APITestCase):
         response = VideoUploadView.as_view()(request)
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        mock_task.assert_called_once()
-        # Get the analysis_id from the call
-        analysis_id = mock_task.call_args[0][0]
-        self.assertEqual(analysis_id, response.data["analysis_id"])
+        # Verify ML endpoint was called
+        mock_requests_post.assert_called_once()
+        # Check that analysis was created with results
+        analysis_id = response.data["analysis_id"]
+        analysis = VideoAnalysis.objects.get(id=analysis_id)
+        self.assertEqual(analysis.status, VideoAnalysis.Status.COMPLETED)
+        self.assertEqual(analysis.analysis_results.count(), 2)
+
+    @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
+    @patch(
+        "videoprocessor.services.video_file_manager.VideoFileManager.extract_and_save_first_frame"
+    )
+    @patch("requests.post")
+    def test_upload_handles_ml_service_error(self, mock_requests_post, mock_extract):
+        """Should mark analysis as FAILED when ML service returns error."""
+        mock_extract.return_value = ("test_thumb.jpg", ContentFile(b"fake thumbnail"))
+
+        # Mock ML endpoint error response
+        mock_response = MagicMock()
+        mock_response.status_code = 500
+        mock_response.text = "Internal Server Error"
+        mock_requests_post.return_value = mock_response
+
+        video_file = self._create_video_file()
+
+        request = self.factory.post(
+            reverse("videoprocessor:video-upload"),
+            {"description": "test", "video": video_file},
+            format="multipart",
+        )
+        force_authenticate(request, user=self.user)
+
+        response = VideoUploadView.as_view()(request)
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        # Check that analysis was marked as FAILED
+        analysis_id = response.data["analysis_id"]
+        analysis = VideoAnalysis.objects.get(id=analysis_id)
+        self.assertEqual(analysis.status, VideoAnalysis.Status.FAILED)
+        self.assertIsNotNone(analysis.error_message)
+
+    @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
+    @patch(
+        "videoprocessor.services.video_file_manager.VideoFileManager.extract_and_save_first_frame"
+    )
+    @patch("requests.post")
+    def test_upload_handles_ml_service_request_exception(
+        self, mock_requests_post, mock_extract
+    ):
+        """Should mark analysis as FAILED when ML service request fails."""
+        mock_extract.return_value = ("test_thumb.jpg", ContentFile(b"fake thumbnail"))
+
+        # Mock request exception
+        mock_requests_post.side_effect = Exception("Network error")
+
+        video_file = self._create_video_file()
+
+        request = self.factory.post(
+            reverse("videoprocessor:video-upload"),
+            {"description": "test", "video": video_file},
+            format="multipart",
+        )
+        force_authenticate(request, user=self.user)
+
+        response = VideoUploadView.as_view()(request)
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        # Check that analysis was marked as FAILED
+        analysis_id = response.data["analysis_id"]
+        analysis = VideoAnalysis.objects.get(id=analysis_id)
+        self.assertEqual(analysis.status, VideoAnalysis.Status.FAILED)
+        self.assertIsNotNone(analysis.error_message)
